添加了一些初始化数据，以及综合评分定时任务

+++ scc_ec/src/main/java/com/baosight/scc/ec/task/DimensionRateOnJob.java
@@ -51,19 +51,20 @@ public class DimensionRateOnJob {
     private void test(){
         logger.info("=====================调度任务启动成功...正在查询...=======================");
     //    String sql = "select * from T_ec_dimension_rate";
-        String sql = "select * from T_ec_sellerCredit";
+        String sql = "select * from T_ec_dimension_rate";
+        logger.info("============================================em="+em);
+        logger.info("============================================query="+em.createNativeQuery(sql));
         List result = em.createNativeQuery(sql).getResultList();
-        for (Iterator iterator = result.iterator();iterator.hasNext();) {
-            Object[] values = (Object[])iterator.next();
-            logger.info("================================id=" + values[0]);
-            logger.info("================================user_id=" + values[1]);
-            logger.info("================================type=" + values[2]);
-            logger.info("================================weekCount=" + values[3]);
-            logger.info("================================oneMonthCount=" + values[4]);
-            logger.info("================================sixMonthCount=" + values[5]);
-            logger.info("================================sixMonthBeforeCount=" + values[6]);
-            logger.info("================================total=" + values[7]);
-            logger.info("================================createTime=" + values[8]);
+        logger.info("============================================result="+result);
+        logger.info("=======================================size="+result.size());
+        if (result.size() > 0){
+            for (Iterator iterator = result.iterator();iterator.hasNext();) {
+                Object[] values = (Object[])iterator.next();
+                logger.info("================================id=" + values[0]);
+                logger.info("================================user_id=" + values[1]);
+                logger.info("================================type=" + values[2]);
+                logger.info("================================weekCount=" + values[3]);
+            }
         }
     }
 
@@ -72,7 +73,7 @@ public class DimensionRateOnJob {
      */
     private void deleteDimensionRates(){
         logger.info("================================删除调度任务启动成功，正在删除T_ec_dimension_rate表中数据...====================================");
-        String deleteSql = "delete from T_ec_dimension_rate";
+        String deleteSql = "delete from T_ec_dimension_rate a where exists (select  id from T_ec_Comment b where a.user_id = b.user_id)";
         Query query = em.createNativeQuery(deleteSql);
         query.executeUpdate();
     }
@@ -129,9 +130,81 @@ public class DimensionRateOnJob {
      */
     private void deleteSellerCredits(){
         logger.info("=====================================调度任务启动成功，正在删除数据表T_ec_sellerCredit中的数据 ... =====================================================");
-        String sql = "delete from T_ec_sellerCredit";
+        String sql = "delete from T_ec_sellerCredit a where exists (select  id from T_ec_comment b where a.user_id=b.user_id)";
         Query query = em.createNativeQuery(sql);
         query.executeUpdate();
     }
 
+    //每晚1:50分执行用户综合评分
+    @Scheduled(cron = "0 50 1  * * ? ")
+    public void addUserCompositeScore(){
+        logger.info("===================调度任务启动成功，正在往T_ec_Composite_Score数据表插入数据 ... ===================================");
+        String sql = "insert into T_ec_Composite_Score(user_id,score,createdTime) ";
+        String s1 = "select user_id,sum(a.attitude+a.deliverySpeed+a.satisfied)/3 as score,current timestamp as createdTime " +
+                " from T_ec_dimension_rate a group by a.user_id,createdTime";
+        sql = sql + s1;
+        //清除数据
+        deleteUserCompositeScore();
+        Query query = em.createNativeQuery(sql);
+        query.executeUpdate();
+    }
+
+    /*
+    清空后表数据
+     */
+    public void deleteUserCompositeScore(){
+        logger.info("=====================================调度任务启动成功，正在删除数据表T_ec_Composite_Score中的数据 ... =====================================================");
+        String sql = "delete from T_ec_Composite_Score a where exists(select id from T_ec_dimension_rate b where a.user_id=b.user_id)";
+        Query query = em.createNativeQuery(sql);
+        query.executeUpdate();
+    }
+    /************************************初始化数据************************************/
+    //每晚0:00执行
+    @Scheduled(cron = "0 0 0  * * ? ")
+    public void initDimensions(){
+        logger.info("===================调度任务启动成功，正在往T_ec_dimension_rate数据表插入数据 ... ===================================");
+        String sql = "insert into T_ec_dimension_rate(user_id,attitude,deliverySpeed,satisfied,createdTime） ";
+        String s1 ="select id,0 as attitude,0 as deliverySpeed,0 as satisfied,current timestamp as createdTime " +
+                " from T_ec_EcUser a where not exists(select id from T_ec_dimension_rate b where a.id=b.user_id)";
+        sql = sql + s1;
+        Query query = em.createNativeQuery(sql);
+        query.executeUpdate();
+    }
+
+    //每晚0:10执行
+    @Scheduled(cron = "0 10 0  * * ? ")
+    public void initSellerCredits(){
+        logger.info("===================调度任务启动成功，正在往T_ec_sellerCredit数据表插入数据 ... ===================================");
+        String sql = "insert into T_ec_sellerCredit(createdTime,user_id,type,weekCount,oneMonthCount,sixMonthCount,sixMonthBeforeCount,total)  ";
+
+        String s1 = "select current timestamp as createdTime,id as user_id,'好评' as type,0 as weekCount,0 as oneMonthCount,0 as sixMonthCount,0 as sixMonthBefore,0 as total " +
+                " from T_ec_EcUser a where not exists(select id from T_ec_sellerCredit b where a.id = b.user_id)  " +
+                "union  " +
+                "select current timestamp as createdTime,id as user_id,'中评' as type,0 as weekCount,0 as oneMonthCount,0 as sixMonthCount,0 as sixMonthBefore,0 as total " +
+                " from T_ec_EcUser a where not exists(select id from T_ec_sellerCredit b where a.id = b.user_id)  " +
+                "union  " +
+                "select current timestamp as createdTime,id as user_id,'差评' as type,0 as weekCount,0 as oneMonthCount,0 as sixMonthCount,0 as sixMonthBefore,0 as total " +
+                " from T_ec_EcUser a where not exists(select id from T_ec_sellerCredit b where a.id = b.user_id)  ";
+
+        String s2 = "select current timestamp as createTimed, a.createdBy,a.type,sum(a.weekCount) as weekCount,sum(a.oneMonthCount) as oneMonthCount,sum(a.sixMonthCount) as sixMonthCount," +
+                "sum(a.sixMonthBeforeCount) as sixMonthBeforeCount,sum(a.total) as total " +
+                " from ("+s1+") a group by type,createdBy,createTimed";
+        sql = sql + s2 ;
+        Query query = em.createNativeQuery(sql);
+        query.executeUpdate();
+    }
+
+
+    //每晚0:50分执行用户综合评分初始化
+    @Scheduled(cron = "0 50 0  * * ? ")
+    public void initUserCompositeScore(){
+        logger.info("===================调度任务启动成功，正在往T_ec_Composite_Score数据表插入数据 ... ===================================");
+        String sql = "insert into T_ec_Composite_Score(user_id,score,createdTime) ";
+        //查询user_id不在T_ec_Composite_Score表中数据，并赋值
+        String s1 = "select id,0 as score,current timestamp as createdTime from T_ec_EcUser a where not exists(" +
+                "select id from T_ec_Composite_Score b where a.id=b.user_id) ";
+        sql = sql + s1;
+        Query query = em.createNativeQuery(sql);
+        query.executeUpdate();
+    }
 }
